% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penfmbr.R
\name{chisq_fmb}
\alias{chisq_fmb}
\title{chisq_fmb}
\usage{
chisq_fmb(object)
}
\arguments{
\item{object}{The fitted fmb model}
}
\value{
A list containing the chi square test statistic and the p value
of all pricing errors jointly being zero
}
\description{
chisq_fmb
}
