% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penfmbr.R
\name{fmb}
\alias{fmb}
\title{fmb}
\usage{
fmb(factors, returns, intercept = TRUE)
}
\arguments{
\item{factors}{The factors}

\item{returns}{The returns}

\item{intercept, }{default TRUE: fit a constant in the second stage}
}
\value{
An `fmb` instance
}
\description{
fmb
}
