% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penfmbr.R
\name{penfmb}
\alias{penfmb}
\title{penfmb}
\usage{
penfmb(factors, returns, alpha = 4, nboot = 100, nblocks = 5)
}
\arguments{
\item{factors}{The factors}

\item{returns}{The returns}

\item{alpha}{The penalty parameter, default 4}

\item{nboot}{Number of bootstrap replications for the shrinkage rate, default 100}

\item{nblocks}{Number of blocks in the stationary bootstrap, default 5}
}
\value{
A `penfmb` instance
}
\description{
penfmb
}
