% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penfmbr.R
\docType{package}
\name{penfmbr}
\alias{penfmbr}
\alias{penfmbr-package}
\title{penfmbr: Plain FamaMacBeth and Penalized FamaMacBeth estimators.}
\description{
For the penalized FamaMacBeth, the level tuning parameter is set to the
average of the residual standard deviation from the first stage. The
dimensionalty paramter is 4 by default, and individual penalty weights are
calculated as the Fisher transform of partial correlations.
}
\details{
`returns` and `factors` are two containers with equal number of observations
(aligned) that can be coerced to a matrix.
Both estimators return a crudely hacked together sublass of `lm` that plays
nicely with `stargazer`.

`vcov` can return either Shanken or GMM standard errors. For a `penfmb`
instance, the bootstrapped shrinkage rate is returned.
}
